# server/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Likesw-Music API
# during dev, should point to your local machine
host: localhost:19001
...
basePath: /music
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /song/create:
    x-swagger-router-controller: song
    post:
      tags:
        - song
      description: Add song information
      operationId: createSong
      parameters:
        - name: "body"
          in: "body"
          description: Song information
          required: true
          schema:
            $ref: "#/definitions/Song"
      responses:
        "200":
          description: Success
          schema:
            type: "string"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /song/load:
    x-swagger-router-controller: song
    get:
      tags:
        - song
      description: load song information
      operationId: loadSong
      parameters:
        - name: name
          in: query
          description: name
          required: false
          type: "string"
        - name: artist
          in: query
          required: false
          description: artist
          type: "string"
        - name: album
          in: query
          required: false
          description: album
          type: "string"
        - name: tags
          in: query
          required: false
          description: tags
          type: "string"
          
      responses:
        200:
          description: OK
          schema:
            type: object
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /collect/save:
    x-swagger-router-controller: collect
    post:
      tags:
        - collect
      description: save collect
      operationId: collectSong
      parameters:
        - name: body
          in: body
          description: info
          required: true
          schema:
            type: string
            format: json
      responses:
        200:
          description: OK
          schema:
            type: string
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  Song:
    properties:
      name:
        type: "string"
      artist:
        type: "string"
      album:
        type: "string"
      tags:
        type: "string"
      image:
        type: "string"
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string